+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/KafkaCluster.java
@@ -15,6 +15,7 @@ import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.LocalObjectReference;
 import io.fabric8.kubernetes.api.model.PersistentVolumeClaim;
 import io.fabric8.kubernetes.api.model.Secret;
+import io.fabric8.kubernetes.api.model.SecretBuilder;
 import io.fabric8.kubernetes.api.model.SecurityContext;
 import io.fabric8.kubernetes.api.model.Service;
 import io.fabric8.kubernetes.api.model.ServicePort;
@@ -1314,9 +1315,11 @@ public class KafkaCluster extends AbstractModel {
     /**
      * Generate the Secret containing the username and password to secure the jmx port on the kafka brokers
      *
+     * @param jmxSecretAnnotations The additional annotations of the {@code Secret} created.
+     * @param jmxSecretLabels The additional labels of the {@code Secret} created.
      * @return The generated Secret
      */
-    public Secret generateJmxSecret() {
+    public Secret generateJmxSecret(Map<String, String> jmxSecretAnnotations, Map<String, String> jmxSecretLabels) {
         Map<String, String> data = new HashMap<>(2);
         String[] keys = {SECRET_JMX_USERNAME_KEY, SECRET_JMX_PASSWORD_KEY};
         PasswordGenerator passwordGenerator = new PasswordGenerator(16);
@@ -1324,7 +1327,17 @@ public class KafkaCluster extends AbstractModel {
             data.put(key, Base64.getEncoder().encodeToString(passwordGenerator.generate().getBytes(StandardCharsets.US_ASCII)));
         }
 
-        return createSecret(KafkaCluster.jmxSecretName(cluster), data);
+        return new SecretBuilder()
+                .withNewMetadata()
+                .withName(KafkaCluster.jmxSecretName(cluster))
+                .withOwnerReferences(createOwnerReference())
+                .withNamespace(namespace)
+                .withLabels(Util.mergeLabelsOrAnnotations(labels.toMap(), jmxSecretLabels))
+                .withAnnotations(jmxSecretAnnotations)
+                .endMetadata()
+                .withType("Opaque")
+                .withData(data)
+                .build();
     }
 
     private List<ContainerPort> getContainerPortList() {