+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/KafkaAssemblyOperator.java
@@ -1230,11 +1230,24 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
         }
 
         Future<ReconciliationState> zkJmxSecret() {
+            Map<String, String> jmxSecretLabels = emptyMap();
+            Map<String, String> jmxSecretAnnotations = emptyMap();
+
+            if (kafkaAssembly.getSpec().getZookeeper() != null
+                    && kafkaAssembly.getSpec().getZookeeper().getTemplate() != null
+                    && kafkaAssembly.getSpec().getZookeeper().getTemplate().getJmxSecret() != null
+                    && kafkaAssembly.getSpec().getZookeeper().getTemplate().getJmxSecret().getMetadata() != null) {
+                jmxSecretAnnotations = kafkaAssembly.getSpec().getZookeeper().getTemplate().getJmxSecret().getMetadata().getAnnotations();
+                jmxSecretLabels = kafkaAssembly.getSpec().getZookeeper().getTemplate().getJmxSecret().getMetadata().getLabels();
+            }
             if (zkCluster.isJmxAuthenticated()) {
                 Future<Secret> secretFuture = secretOperations.getAsync(namespace, ZookeeperCluster.jmxSecretName(name));
+                Map<String, String> finalJmxSecretAnnotations = jmxSecretAnnotations;
+                Map<String, String> finalJmxSecretLabels = jmxSecretLabels;
                 return secretFuture.compose(secret -> {
                     if (secret == null) {
-                        return withVoid(secretOperations.reconcile(reconciliation, namespace, ZookeeperCluster.jmxSecretName(name), zkCluster.generateJmxSecret()));
+                        return withVoid(secretOperations.reconcile(reconciliation, namespace, ZookeeperCluster.jmxSecretName(name),
+                                zkCluster.generateJmxSecret(finalJmxSecretAnnotations, finalJmxSecretLabels)));
                     }
                     return withVoid(Future.succeededFuture(ReconcileResult.noop(secret)));
                 });
@@ -2464,12 +2477,24 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
         }
 
         Future<ReconciliationState> kafkaJmxSecret() {
+            Map<String, String> jmxSecretLabels = emptyMap();
+            Map<String, String> jmxSecretAnnotations = emptyMap();
+
+            if (kafkaAssembly.getSpec().getKafka() != null
+                    && kafkaAssembly.getSpec().getKafka().getTemplate() != null
+                    && kafkaAssembly.getSpec().getKafka().getTemplate().getJmxSecret() != null
+                    && kafkaAssembly.getSpec().getKafka().getTemplate().getJmxSecret().getMetadata() != null) {
+                jmxSecretAnnotations = kafkaAssembly.getSpec().getKafka().getTemplate().getJmxSecret().getMetadata().getAnnotations();
+                jmxSecretLabels = kafkaAssembly.getSpec().getKafka().getTemplate().getJmxSecret().getMetadata().getLabels();
+            }
             if (kafkaCluster.isJmxAuthenticated()) {
                 Future<Secret> secretFuture = secretOperations.getAsync(namespace, KafkaCluster.jmxSecretName(name));
+                Map<String, String> finalJmxSecretAnnotations = jmxSecretAnnotations;
+                Map<String, String> finalJmxSecretLabels = jmxSecretLabels;
                 return secretFuture.compose(res -> {
                     if (res == null) {
                         return withVoid(secretOperations.reconcile(reconciliation, namespace, KafkaCluster.jmxSecretName(name),
-                                kafkaCluster.generateJmxSecret()));
+                                kafkaCluster.generateJmxSecret(finalJmxSecretAnnotations, finalJmxSecretLabels)));
                     }
                     return withVoid(Future.succeededFuture(this));
                 });