+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Update the Open Policy Agent Authorizer to version [1.1.0](https://github.com/Bisnode/opa-kafka-plugin/releases/tag/v1.1.0)
 * Expose JMX port on Zookeeper nodes via a headless service.
+* Allow configuring labels and annotations for JMX authentication secrets
 
 ### Changes, deprecations and removals